requirements.txt
pymongo>=4.6.0
mongomock>=0.4.1
psycopg2-binary>=2.9.9
mysql-connector-python>=8.3.0
tabulate>=0.9.0
pytest>=7.4.0
pytest-cov>=4.1.0
